moved common test helper methods to separte package
